requirements.txt
Flask==2.0.1
Flask-Blueprint==0.0.3
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
Jinja2==3.0.1
neo4j==4.3.3
python-dotenv==0.18.0